// src/api/store/cart/package-items/validators.ts
import { z } from "@medusajs/framework/zod"

/**
 * Body for POST /store/cart/package-items
 */
export const StoreAddPackageItemsBody = z.object({
  cart_id: z.string().optional(),
  package_id: z.string().optional(),
  package_date: z.string().optional(),
  adults: z.coerce.number().optional(),
  children: z.coerce.number().optional(),
  infants: z.coerce.number().optional(),
  customer: z
    .object({
      name: z.string(),
      email: z.string().email().optional(),
      phone: z.string().optional(),
      formId: z.any().optional(),
    })
    .optional(),
  items: z
    .array(
      z.object({
        variant_id: z.string(),
        quantity: z.coerce.number().optional(),
        unit_price: z.coerce.number().optional(),
        thumbnail: z.string().optional(),
        metadata: z.record(z.string(), z.any()).optional(),
      })
    )
    .optional(),
  })

export type StoreAddPackageItemsBodyType = z.infer<typeof StoreAddPackageItemsBody>